move around plt helpers
