+++ spiketools/plts/style.py
@@ -5,54 +5,11 @@ from functools import wraps
 import matplotlib.pyplot as plt
 
 from spiketools.utils.checks import check_list_options
+from spiketools.plts.utils import check_ax, get_kwargs, get_attr_kwargs
 from spiketools.plts.settings import SET_KWARGS, OTHER_KWARGS
 
 ###################################################################################################
 ###################################################################################################
-
-def get_kwargs(kwargs, select):
-    """Get keyword arguments.
-
-    Parameters
-    ----------
-    kwargs : dict
-        Keyword arguments to extract from.
-    select : list of str
-        The arguments to extract.
-
-    Returns
-    -------
-    setters : dict
-        Selected keyword arguments.
-    """
-
-    setters = {arg : kwargs.pop(arg, None) for arg in select}
-    setters = {arg : value for arg, value in setters.items() if value is not None}
-
-    return setters
-
-
-def get_attr_kwargs(kwargs, attr):
-    """Get keyword arguments related to a particular attribute.
-
-    Parameters
-    ----------
-    kwargs : dict
-        Plotting related keyword arguments.
-    attr : str
-        The attribute to select related arguments.
-
-    Returns
-    -------
-    attr_kwargs : dict
-        Selected keyword arguments, related to the given attribute.
-    """
-
-    labels = [key for key in kwargs.keys() if attr in key]
-    attr_kwargs = {label.split('_')[1] : kwargs.pop(label) for label in labels}
-
-    return attr_kwargs
-
 
 def set_plt_kwargs(func):
     """Collects and then sets plot kwargs that can be applied with 'set'."""
@@ -92,8 +49,6 @@ def drop_spines(sides, ax=None):
         Axis object to update.
         If not provided, takes the current axis.
     """
-    # TEMP: avoid circular import
-    from spiketools.plts.utils import check_ax
 
     ax = check_ax(ax, return_current=True)
 
@@ -101,3 +56,29 @@ def drop_spines(sides, ax=None):
     check_list_options(sides, 'sides', ['left', 'right', 'top', 'bottom'])
     for side in sides:
         ax.spines[side].set_visible(False)
+
+
+def invert_axes(invert, ax=None):
+    """Invert plot axes.
+
+    Parameters
+    ----------
+    invert : {'x', 'y', 'both'} or None
+        How to invert the plot axes, inverting the x, y, or both axes.
+    ax : Axes, optional
+        Axis object to update.
+        If not provided, takes the current axis.
+
+    Notes
+    -----
+    Note that for a 2d array, inverting axes is equivalent to flipping the data, specifically:
+        Flipping up/down is equivalent to inverting the y-axis.
+        Flipping left/right is equivalent to inverting the x-axis.
+    """
+
+    ax = check_ax(ax, return_current=True)
+
+    if invert in ['x', 'both']:
+        ax.invert_xaxis()
+    if invert in ['y', 'both']:
+        ax.invert_yaxis()
